linker_array: add linker arrays
Add "linker array" header file.  This file defines macros that allow
building arrays across compilation units, using the linker.  Entries are
placed in special sections that are placed together by the linker,
allowing them to be treated as arrays.  They are placed in read-only
memory.

This will allow special sections, like .user and .clocks, to be removed
in favor of a general interface.

Change-Id: I0c124cad113eb9fabf8e76119cb895bba9c63b3e