dev: ADC class
Add an ADC class.  A linker array contains all available ADC drivers and
their compatibility strings.  When getting an ADC object for a GPIO,
look through this list and the device tree for a device and driver
capable of creating the object.

The class itself is based on a GPIO object, and simply provides a raw
data reading method.

Change-Id: Ie6ee2155e355a2bbf88ddebf1f8ee33588a30ec1