stm32f4: ADC driver
Standard ADC driver for the STM32F4 series.  Decodes the channel GPIOs
from the FDT, and keeps track of allocated channels.  When a new ADC is
requested from a GPIO, the driver finds a compatible, free channel, and
prepares it for scanning.

The actual read_raw() method actually ignores the sequence setup and
sets the channel to read as the first channel in a one channel sequence.
This one channel is converted and read within the function.

Eventually, DMA should be used to copy converted data, allowing the
entire channel sequence to be converted.

Change-Id: Ifccf751f84126c66e6ab51d4d972a04c14620221