kernel: add reentrant mutexes
Implement reentrant mutexes entirely in userspace.  They allow locks to
be acquired multiple times by the same task, so long as they are
released the same number of times.

The locks store the task that has them acquired, allowing that task to
reacquire, and forcing others to perform an actual acquire.

Change-Id: Ic0eb412145e6d739a27e109db81ecdd03eaebbbc