kernel: add WARN_ON() and PANIC_ON()
Add macros to output a warning, and in the case of PANIC_ON(), panic the
system.

Currently, the output is very basic, simply printing the assertion that
failed and its location in the OS.

These should be used throughout the kernel to make assertions about
internal state, where an error cannot be returned to the user.

Change-Id: I7c8c11befb01b356bd6b4aa773389c8b6b912b9f