kernel: correct list_add() argument order
In list_add(), the item to add comes first, followed by the list to add
to.  collection_add() got the argument order backwards.

In the collections interface, this manifested itself as
collection_iter() returning the correct value, followed by
collection_next() returning a bad value.  The value was in memory, so it
could be dereferenced, but was not the correct type.

As a result, only the first item in a collection could be retrieved.

Change-Id: I6454fcd07abfaaf2690f413fcad63f3e9f005935