armv7m: switch to PSP at reset
Switch to using the PSP immediately on reset.  This allows extra logic
in the SVC handler detecting SVCs coming from the MSP to be removed.

This avoids a bug where save_context would save the caller's context on
the PSP, which was still 0x0, potentially causing a bus fault.

The switch to interrupt context will switch the core to the MSP.  To
keep the interrupt context from cobblering the thread mode stack, leave
a 512 byte buffer between the initial MSP and PSP.

Change-Id: I3ce79b0a295060e5d1d7b3f347521f14518ad56b