build: build with arch-level linker script
Generalize the linker scripts and provide a single linker script for
each architecture.  The section VMA and LMA regions are specified by a
new set of Kconfig options.

The linker memory regions section is automatically generated from a new
/memory node in the device tree.  This node describes the ROM, RAM,
flash, etc regions of memory and is parsed at build time to build LD's
MEMORY section for the linker script.

Along the way, some unused symbols are removed and the linker scripts
are generally cleaned up.

Change-Id: I3f65cdf764286c9287a451a3bec144ab07fa43b8