autopilot: initial sample program
Experiment with a partitioned autopilot program.  The program is split
into two tasks:

A magnetometer reading task handles communication with the magnetometer,
running every 20ms and writing the data to a locked global.

A "controller" task reads the mag data from the locked global, and
prints the value, every 10ms.

This method works, but the IPC mechanism could use work, as it is
rudimentary, manual, and rather annoying.

Other areas for improvement:
    * new_task() should take a period in (micro)seconds, not ticks
    * A blocking task state would be useful for handling things like the
      reading the baro, which should be high priority, but can't be
      because they spend too long waiting.

Change-Id: Ie7a189d622b23a033c86d8d00b7ea6f537ee9561