sched: new_task() takes period in microseconds
Instead of taking task period as platform and configuration specific
ticks in new_task(), take the period in microseconds.  new_task()
converts this value to ticks, which are still used by the remainder of
the scheduler.

new_task() will round up to the nearest number of ticks that can be
achieved by the scheduler.  For example, if the tick period is 250us,
and a task with a period of 200us is created, it will actually have a
period of 250us.

Change-Id: Ie985eb74b01e42454fc73deb78ddafea2c7a46c4