dev: character device abstraction
Introduce a character device obj type.  This type will act as a wrapper
for more concrete character device types, such as UART or USB CDC.  It
implements a generic read/write interface, without hardware-specific
configuration methods (such as setting baud rate).

The goal with this type is to replace the resource system.  printf() and
friends would write to character devices, which could be UART, USB, etc.

Change-Id: I48ac5ecaf8a586a1e281dc5f15ca9d4f1f65c387