dev: resource char_device implementation
To ease the transition to character devices, implement a character
device layer on top of the resource system.  Eventually, all uses of
resources can be replaced with character devices based on concrete
peripheral objs, but for now, this allows the transition to begin while
some resources are still in use.

Update read()/write()/swrite() to use the char_device internally.
Currently, this is very inefficient, as it requires a malloc() on each
call.  The next step is to make read()/write()/swrite() take
char_devices directly, rather than building them each time.

Additionally, resource_write() never makes use of the resource's
swriter() method, so USB output is noticably slower.

Change-Id: Ieeccc553f3bd13b33077c4f6779729eea36b19d9