stm32f4: uart class
Implement the UART class for the STM32F4.  This enables the use of UARTs
besides UART1.  Note that UART1 is still used for standard resources, so
it should not be used through this class until that use is removed.

The UART peripheral clock speed is still hard-coded, so it likely will
incorrectly set the baud rate for some UART peripherals.

Change-Id: I10fe87ec7e0f70b7f8352dcd2dba3f5350de425a