dev: cast to char
Implement a generic char_device_cast() function, which takes an obj of
indeterminate type, and returns a char_device which utilizes the base
obj.  This makes it possible to get a device from device_get(), cast it
to a char_device, and use it generically.  This is provided by
char_device_get().

This is implemented with a linker list of "conversion" functions.  These
functions take an obj of definite type and return a char_device for
them.  char_device_cast() simply looks for a conversion function
matching the type of the obj passed.

Additionally, the reference behavior of character devices is better
defined.  They must have their own reference, meaning that casting
functions should perform an obj_get() on the base obj.  If the base obj
is no longer needed, it should be obj_put() after being cast.

Change-Id: I27f2af5a50dc1cde0af737839a01cfaa93985b72