stm32f4: USB char_device wrapper
Replace the STM32F4 USB resource wrapper with a char_device wrapper.  It
is not a true modernized, or FDT-based driver, but simply a char_device
wrapper around the existing USB driver.

The driver is registered as the device "stm32f4-static-usb", which can
be used to make it stdout.

Change-Id: I7738197e191c771cbeb085ccbf337b841db94204