stm32f4: set and get SPI clock rate
Add internal functions for getting and setting the SPI clock rate.
These could be used for get/set clock methods on the SPI obj.

When setting the clock, the divisor is rounded up, hence the clock rate
will be rounded down to the nearest possible value.

Default the SPI peripheral to 1MHz.

Change-Id: I5ea28a1b5a9446922422a62bd1f9bae3a83db8aa