sched: avoid double adding periodic tasks to runnable task list
When a periodic task has zero ticks remaining until it should run, it
must be added to the runnable task list for scheduling.

However, if the task is already in the list, it must not be added again,
as that would corrupt the list.  The existing check only verified that
the task was not running.  However, it is possible that the task was
placed in the runnable task list one period ago and was never scheduled,
so it is still not running.

Directly check that the task is not in the runnable task list before
adding it.

Change-Id: Id110b604977eff3188420742f18c649801bf231a