msp432: gpio driver
Add a GPIO driver the for MSP432, which should be fully capable for
supporting other driver's needs in the OS.

Currently, the driver looks up the GPIO register base address from the
first compatible device tree node.  Ideally, it would perform the lookup
based on a phandle reference to a GPIO peripheral, but the GPIO API does
not currently pass enough information through for this to be possible.

Change-Id: I3feada3bc70f928e11f2b7062e11f9d043b738a2