armv7m: armv7a: strex status destination register must be unique
The ARMv7-M and ARMv7-A Architecture Reference Manuals state for the
strex instruction:

if d == n || d == t then UNPREDICTABLE;

Where d is the status destination register, n is the source register,
and t is the base register.

I have not noticed any unexpected behavior with this instruction, I
simply noticed that gas enforces this constraint with strexb and strexh
instructions, but not strex instructions.

In inline assembly blocks, mark the status destination register as
'output only'.

Change-Id: I6b7a48c712f3e69f00cfc32af55597ecb58a7901